--- application.properties
@@ -1,1 +1,0 @@
-app.grails.version=2.0.0
