+++ gradle.properties
@@ -1,0 +1,2 @@
+grailsVersion=3.1.1
+gradleWrapperVersion=2.9
