--- grails-app/conf/Config.groovy
@@ -1,12 +1,0 @@
-log4j = {
-	error 'org.codehaus.groovy.grails',
-	      'org.springframework',
-	      'org.hibernate',
-	      'net.sf.ehcache.hibernate'
-	info   'grails.plugin.hibernatehijacker'
-}
-
-grails.doc.authors="Kim A. Betti"
-grails.doc.license="Apache 2.0"
-grails.doc.copyright=""
-grails.doc.footer="Have a nice day!"
