--- grails-app/conf/DataSource.groovy
@@ -1,14 +1,0 @@
-dataSource {
-	pooled = true
-	driverClassName = 'org.h2.Driver'
-	username = 'sa'
-	password = ''
-	dbCreate = 'update'
-	url = 'jdbc:h2:mem:testDb'
-}
-
-hibernate {
-	cache.use_second_level_cache = false
-	cache.use_query_cache = false
-	cache.provider_class = 'org.hibernate.cache.EhCacheProvider'
-}
