+++ grails-app/controllers/hibernatehijacker/UrlMappings.groovy
@@ -1,0 +1,16 @@
+package hibernatehijacker
+
+class UrlMappings {
+
+    static mappings = {
+        "/$controller/$action?/$id?(.$format)?"{
+            constraints {
+                // apply constraints here
+            }
+        }
+
+        "/"(view:"/index")
+        "500"(view:'/error')
+        "404"(view:'/notFound')
+    }
+}
