+++ src/integration-test/groovy/grails/plugin/hibernatehijacker/demo/BookFunctionalSpec.groovy
