+++ src/main/groovy/grails/plugin/hibernatehijacker/indexdsl/HibernateIndexBuilder.groovy
