+++ src/main/groovy/grails/plugin/hibernatehijacker/template/HibernateTemplates.groovy
