+++ src/main/java/grails/plugin/hibernatehijacker/exception/HibernateHijackerException.java
