+++ src/test/groovy/grails/plugin/hibernatehijacker/hibernate/events/HibernateEventPropertyUpdaterSpec.groovy
@@ -1,0 +1,15 @@
+package grails.plugin.hibernatehijacker.hibernate.events
+
+import org.hibernate.tuple.entity.EntityMetamodel;
+
+/**
+ * 
+ * @author Kim A. Betti
+ */
+class HibernateEventPropertyUpdaterSpec {
+
+    // Turns out that the huge object graphs in the
+    // Hibernate API isn't easy to test :P 
+    // (another day..)
+
+}
